Add support for uninitialized/unknown values in FSA
Added the DataValue class to represent values that are uninitialized
(and thus unknown)
This is used when developing the FSA from a data stream. Each state
depends on a number of variables, as defined by the user in the DSL.
However, until the first record of a execution of each event, the value
for that event variable is unknown, and thus needs a way to be
represented in the FSA.
Implemented the use of DataValue objects rather than direct integers.